Finished connection to cloudant.
Moved core code to poll module and cleaned up slack poll. Removed the tests since they no longer are useful. Added the manifest file that is used to deploy the app.
